Include non-apps on the "ownership by team" page
The 'apps.html' page is the source of truth as to whether a given
repo is an application or not.
The 'by-team.html' page is the source of truth as to whether a
given repository is owned by a given team. It does not need to be
restricted to just applications. Indeed, there are plenty of repos
which require ownership/maintenance/Dependabot effort, but which
aren't classic 'applications', e.g. govuk-docker.