Remove unnecessary config from repos.yml
It shouldn't be necessary to supply `dashboard_url`, `deploy_url`
and `sentry_url` as `false` if they're not appropriate for the
given repo - they should simply be omitted.

Not providing them used to raise an exception, but now our
rendering code is more defensive, so it looks like it's all working
as it should.