Remove section `UNKNOWN: INTERNAL ERROR`
[Trello](https://trello.com/c/AGdifoDb/250-remove-fabric-and-replace-references-in-docs)

I was setting out to just remove the reference to fab here, but the
archived fab repo seems to have had statsd commands purposely removed
from it before we archived the repo.

Quote from an [old PR][fabric_scripts#291]:

> I don't think this has been used in a while, and I also think there
> are better approaches now of addressing the problem it's trying to
> solve.

> I think this was about getting rid of UNKNOWN Icinga checks, but it's
> better just to acknowledge the failing check, or change the Graphite
> query if the data just hasn't arrived yet.

If that's the case then likely better to just clear up this entire
section.

[fabric_scripts#291](https://github.com/alphagov/fabric-scripts/pull/291)