Update docs concerning detach and terminate in AWS
This commit does the following:

* consolidates the removal of instances section from auto-scaling-groups
into common-aws-tasks-for-2nd-line-support, as the former documentation
is geared towards scaling up, or down, rather than detaching instances.

* Removes 'reboot' as an example of when you would want to detach an
instance as:

a) the reboot docs linked don't mention it

b) for a reboot we probably want to set the instance to 'Standby' as we
expect the instance to be put back in service, rather than terminated.
This is easier using 'Standby' as the instance capacity is automatically
adjusted.

* Add specific section around terminating an instance, which follows as
a potential action after detaching an instance.

* Uses AWS terminology (detach vs remove).

* Point to AWS docs, rather than duplicate it in our own.

* Removes Users section as everyone with production access has
permission to assume the admin role so this information is no longer
useful in common 2nd lines tasks doc.