Update env-sync documentation
I started updating this to reflect that there were now lots of db-admin
machine but took tangents updating other aspects that were out of date.
Most noticeably the diagram was quite unrealistic compared to how the
tool works now (and seemed better replaced with words).

The configuration for the script seemed unnecessary since it's all
provided in the file [1] which is linked to in this doc (and the
information here had fallen out of sync.

[1]: https://github.com/alphagov/govuk-puppet/blob/main/modules/govuk_env_sync/files/govuk_env_sync.sh