Reflect multiple db_admin instances in govuk-in-aws
This updates this document to reflect that there are now multiple
db_admin instances. It also removes the link to administer the RDS
instances since the destination was a very specific restore process and
not general administration.

To be honest, I wimped out here in a nuke of this file. It seems to have
lost it's value now GOV.UK has been migrated to AWS for years (and the
next infrastructure migration has already clocked in over a year of
work). However there still seems to be links to it and bits of things
documented here that aren't elsewhere. Sorry...