Remove GOV.UK Accounts Applications section
This section is currently empty, it used to be populated by the now
archived [govuk-account-manager-prototype][1] and
[govuk-attribute-service-prototype][2].

These were replaced with the GDS Digital Identity Single Sign On
integration, and an enhanced role for [account-api][3] in managing
GOV.UK specific data.

Since then, all applications of a GOV.UK account are intended to be
integrated into other GOV.UK functions, so a separate category no longer
makes sense and can be removed.

[1]: https://github.com/alphagov/govuk-account-manager-prototype/
[2]: https://github.com/alphagov/govuk-attribute-service-prototype/
[3]: https://github.com/alphagov/account-api/