Handle private repos' lack of README gracefully
Requests for the `github_readme` method of a private repo return
`nil`:
https://github.com/alphagov/govuk-developer-docs/blob/7427f6886c0676a483e837824fb7d53a19db6df7/app/app.rb#L208

This is useful as it lets us provide specific error messages:
https://github.com/alphagov/govuk-developer-docs/blob/c4b9aebacb3d2c6f2948a5d96ef1c288fc2cd47b/source/partials/application/_readme.html.erb#L9-L13

However, calling `ExternalDoc.parse(application.readme` in
https://github.com/alphagov/govuk-developer-docs/blob/c4b9aebacb3d2c6f2948a5d96ef1c288fc2cd47b/source/partials/application/_readme.html.erb#L15
meant we were passing `nil` to the HTML::Pipeline's `to_html`
method. Even though we're attempting to be defensive and cast it
with `.to_s`, this was raising a "Cannot modify frozen string"
exception.