Skip loading of proxy pages with an env var
Often I find myself wanting to work on a few pages of the dev docs which
are local to this repository.

It's annoying having to wait for my laptop to download all of GitHub
before middleman will start serving the entirely local content I want to
edit.

Adding this flag gives me a way to run the middleman server and work on
these pages without needing to make requests to GitHub.

Unfortunately, middleman server will still download stuff that's
requested via .erb files when you go to the homepage / apps page etc. I
had a look at disabling all of this, but I'd need to do a big refactor
to make it clean so this small change was the best I could come up with.