Update instructions for reindexing search
In [1], we have added an index selector to the search reindex Jenkins
job.

This means we can remove all references to the two options (one being
the Jenkins job for all indices and the rake task for single indices).
All reindex tasks can now be done with the same Jenkins job.

1: https://github.com/alphagov/govuk-puppet/pull/11548