Update conventions for Jasmine JS testing
This commit updates the Rails conventions manual to reflect the new approach we've taken for running Jasmine tests.

This change has come about because the `jasmine` gem has been deprecated, and it doesn't support Rails 7. We've therefore had to introduce this new approach as an enabler for our applications to run on Rails 7.