Update docs to state GCP access is needed for DNS deploys
This was updated in https://github.com/alphagov/govuk-puppet/pull/11168
to avoid having to store google credentials with administrative
privileges in Jenkins.

The process now involves using the gcloud CLI to get a temporary access
token, and providing this to Jenkins via the API (as we do with AWS
tokens).