Remove Publishing E2E tests from app links
This was a bit confusing how it gave you a link and then warned you it
mightn't work - especially as the links are all conditional.

I doubt anyone has ever got value out of this and it's
implemented so inconsistently (many apps that are part of E2E don't have
this) that I think it's simpler to remove.