Add repos.json and repos.csv
These files allow machine readable lists of all of the GOV.UK repos.

The use case for this is to replace usage of apps.json with repos.json
in govuk-dependencies, so that teams can be notified of dependencies of
utilities they own which are not apps.

There isn't a current use-case for the CSV file, but I thought this was
good to have for consistency.