Expand on docker_management reboot reasons
The docker_management machine runs etcd and is used as a lock by other machines
that wish to perform an unattended reboot. For that reason, we need to be
a bit careful about rebooting it.

The etcd host is specified by the [etcd_endpoints](https://github.com/alphagov/govuk-puppet/blob/main/hieradata_aws/common.yaml#L1544)
configuration.

There is further context available in:
- [puppet-unattended_reboot Github repo](https://github.com/alphagov/puppet-unattended_reboot)
- [a presentation by Matt Bostock (@mattbostock)](https://www.dotconferences.com/2015/06/matt-bostock-coordinating-unattended-reboots-using-a-distributed-mutex)