Add section about Publishing API special case
Confusingly, the duplicate document in Publishing API [^1] isn't
about this special case, but pact testing with GDS API Adapters.

Since the special case applies to Content Store as well, I think
it's better to have it documented centrally here.

I've also swapped the intro references to Publishing API, as it's
confusing to use a special case as the main example.

[^1]: https://github.com/alphagov/publishing-api/blob/dd8dd9232d3cbf33b8945fdd898ebe80d7dcfcf6/docs/pact_testing.md