Retire webpacker aspiration for vaguer modern one
The content in this convention no-longer reflected reality. Rails has
moved away from Webpacker since Rails 7 and thusly, so has our
preference to use it.

We do however aspire to use modern Javascript, but which form we will
take is not clear. However whatever road we take, we will need to
resolve the asset pipeline coupling of GOV.UK Publishing Components,
which is the main limiter for newer apps.

This update tries to condense that into a single sentence.