Rename 'App' to 'Repo'
We've already renamed 'Applications' to 'Repos' to reflect the
increased scope of the class, which now manages repos of both
app and non-app types. The underlying model that Applications (now
Repos) was managing was the App model, but now an App can represent
non-apps, so should be renamed Repo.

We can use the new 'is_app?' method to check if a repo is
considered to be a GOV.UK app, i.e. is hosted somewhere.

I did consider creating a separate class, e.g. Repo, and
retaining 'App' as a subclass of Repo. But this opens a can of
worms: we'd have to consider the 'production_hosted_on'
information inside Repos.rb's `self.all` method (or delegate to
a factory that returns either a Repo or App).

We'd have to consider every usage of App/Repo in the codebase
and consider which of the two it should be, which would make
future changes difficult.

There's also the question of whether we'd want to retain the old
'/apps/*.html' path for Apps, since there'd be more of a concrete
separation between the two. This would make the Dev Docs less
predictable; it'd be harder to know what the overview URL for a
given repo would be, without doing some digging.

All things considered, just renaming 'App' to 'Repo' feels like
the pragmatic move here (and better than leaving it named as a
potentially misleading 'App'). We can always revisit later if we
wanted to do something more fancy.