Rename 'app_name' method to 'repo_name'
This method is used all over govuk-developer-docs, in the context
of both Apps and non-Apps. As part of the work to remove
potentially misleading code from govuk-developer-docs, this has
now been renamed to 'repo_name', which works in every case.

This does mean we lose the ability to override app names where we
want them to differ from the repo names, but in practice, this
was only used in two places:

1) blinkenjs (app_name: govuk-secondline-blinken)
   The management URL may be https://dashboard.heroku.com/apps/govuk-secondline-blinken,
   but other than that, no references to this app name exist.
   We should rename the repo rather than have an override.
2) govuk-content-schemas (app_name: govuk-content-store-examples)
   Not actually the name of the app at all. The Release app calls
   it "GOV.UK Content Schemas". The only reference to this app
   name is the management URL: https://dashboard.heroku.com/apps/govuk-content-store-examples

There are other places where app name overrides are crucial to
the functionality or deployment, e.g. smart-answers vs smartanswers:
https://github.com/alphagov/govuk-puppet/blob/b93271f7e6448a3094f2e39895e93b5470d09477/hieradata_aws/common.yaml#L205-L206

This is not one of those places. By removing the ability to
override, we make the code much simpler.

We're now left with a `repo_name` method that calls the
`github_repo_name` method. We'll tidy that up in the next commit.