Add 'Repos' page and consolidate the 'Apps' page
This creates a new 'repos.html' page, which is intended to be a
catch-all page similar to the 'apps.html' page, listing all known
GOV.UK repositories:

![repos page screenshot](https://user-images.githubusercontent.com/5111927/166255289-07212299-65d5-43f4-8178-0455c14699ff.png)

The previous implementation of the apps.html page was not a great
design; it added many links to the sidebar, which had to be
scrolled to see all the available apps. Better to put this
information on the main page itself, and default to the `layout`
layout, which automatically parses page headings:
https://tdt-documentation.london.cloudapps.digital/configure_project/frontmatter/#layout

Before:

![before](https://user-images.githubusercontent.com/5111927/166255046-8042ab9d-4065-4f02-9479-6d16b86dcd99.png)

After:

![after](https://user-images.githubusercontent.com/5111927/166255030-d79082f0-07ce-4e9b-bfa6-4bfe3a4c992b.png)

Now, all 'grouped' information is available in one place (teams,
types, hosting), and the sidebar is used for an arguably better
purpose: anchor links to the relevant parts of the page.
This also removes the need for a redundant 'by-teams.html' page,
as the info is already captured on the apps.html page.

The result is a simpler codebase and more easily findable
information.