Remove custom sidebar from application layout
The sidebar was very long, listing all active apps - and arguably
is not useful. If you're reading an imported doc from account-api,
why would you want to be able to quickly jump to `mapit`?

The best thing would be the default layout, which would remove
those pointless links and instead show the search bar and a table
of contents of the imported page, giving you an at-a-glance overview
of the page structure and anchor links so that you can quickly jump
to the relevant section.

However, using the standard layout would mean losing our custom
'last updated' banner, and breadcrumb trail. Therefore, specifying
an empty (but not `nil`) sidebar is the best compromise at this stage.
A `nil` sidebar would mean no sidebar gets rendered at all, meaning
the user loses their ability to search from any page, which wouldn't
be ideal.

Before:

![before](https://user-images.githubusercontent.com/5111927/166257912-9f106d4f-4383-4d1b-be90-829aab023a81.png)

After:

![after](https://user-images.githubusercontent.com/5111927/166257893-b360eda1-1c5e-4ce4-bb9a-7f47ec1d0f1a.png)