Move 'imported docs' list into sidebar
The sidebar was very long, listing all active apps - and arguably
is not useful. If you're reading an imported doc from account-api,
why would you want to be able to quickly jump to `mapit`?

The best thing would be the default layout, which would remove
those pointless links and instead show the search bar and a table
of contents of the imported page, giving you an at-a-glance overview
of the page structure and anchor links so that you can quickly jump
to the relevant section.

However, using the standard layout would mean losing our custom
'last updated' banner, and breadcrumb trail. Therefore, specifying
something similar to that - and useful, and contextual - is best.
We can surface all the imported documents in the sidebar. We can
therefore remove them from the main repo overview page.

Note that the menu links have been improved by adding 'rel="noopener"',
which is a [security recommendation](https://developer.mozilla.org/en-US/docs/Web/HTML/Link_types/noopener).

Note also that I considered using the `list` component from
govuk_publishing_components, but it had different styling to the
rest of the Developer Docs menus, and also meant complicating the
template (since all list items had to be stored as ruby variables
so that they could be rendered by a call to
`GovukPublishingComponents.render`). Using hand-written HTML to
match the other menus' HTML made more sense, and was worth the
two lines of CSS as trade-off.

Before:

![before](https://user-images.githubusercontent.com/5111927/166269552-968c9488-73d8-4457-9d5b-b91ac8b256fb.png)

After:

![after](https://user-images.githubusercontent.com/5111927/177963393-8f4f81eb-ee74-47bd-87a9-2113aaf8cbfc.png)

Trello: https://trello.com/c/oUsgzZnk/247-standardise-how-we-represent-repositories-in-the-dev-docs