Rename 'application/*' partials to 'repo/*'
This removes the last element of confusion; we were still referring
to a 'application' template variable, even when rendering a non-app
(i.e. a repo such as a gem).

Many of the aspects of these views _are_ written with applications
in mind, but the rename should now make it clear that we need to
write the templates to handle non-apps too.