Move 'imported docs' list into sidebar
We'd already established that showing a list of all apps in the
sidebar was not very useful, and temporarily removed everything
but the search form from the sidebar.

The traditional layout parses the page contents and surfaces
anchor links in a 'table of contents' style. Something similar to
that - and useful, and contextual - is to surface all the imported
documents in the sidebar. We can therefore remove them from the main
repo overview page.

Before:

![before](https://user-images.githubusercontent.com/5111927/166269552-968c9488-73d8-4457-9d5b-b91ac8b256fb.png)

After:

![after](https://user-images.githubusercontent.com/5111927/166269563-7f0ab5aa-4163-41da-90a3-242480dea33a.png)