Document the databases on GOV.UK
The original scope was just to document the DB Admin machines, but
this assumes a lot of knowledge about the GOV.UK database
infrastructure that is not already neatly consolidated in one
place. So rather than just document the DB Admin machines, this
commit documents GOV.UK's use of databases in general.

Trello: https://trello.com/c/0LHZ1zo7/2840-document-the-dbadmin-infrastructure