Fix out-of-date guidance about custom healthchecks
Content Publisher no longer has a custom healthcheck. I also think
it's more useful to point to the dev docs and there are only three
apps that have these so it's manageable to list them all*.

I've also made each of the custom docs consistent:

- They all link back to this general manual.
- They all link to the healthcheck endpoint in the app.

Note that the "look at the logs" suggestion for Search API is also
repeated directly below, so we're not losing anything.

*Checked in Puppet by searching for "health_check_custom_doc".