Rewrite intro to cover different healthchecks
The previous guidance was out-of-date:

- We now have two types of healthcheck, although only one should
be relevant for this alert [^1].

- Some of the previous examples no longer exist or have changed.

The only apps with a custom healthcheck endpoint appear to be CKAN,
licensing apps and other outliers like Mapit - checked by looking
for "health_check_path" in Puppet. The term "legacy" is a bit vague
but hopefully enough to disuade people from this approach.

[^1]: https://github.com/alphagov/govuk-puppet/pull/11690