Tidy up contents section of the home page
It was unclear what the itention was for this page: the headings
seem arbitrary and ones like "Transition" don't make sense without
(a lot) more context - we have a manual for it [^1].

This removes all links to individual repos as these are now provided
more robustly by the "Repos" and "Apps" pages. I've also removed the
link to Sentry as we have a manual about it [^2].

I'm not sure exactly what should be on this page, but I suspect it's
most useful to someone who's new to GOV.UK and never seen these docs
before, so let's start tailoring it to that use case.

[^1]: https://docs.publishing.service.gov.uk/manual/transition-architecture.html
[^2]: https://docs.publishing.service.gov.uk/manual/sentry.html