Add guidance on how we test GOV.UK
This is the culmination of:

- RFC 128, which was really a new testing strategy [^1].
- Work to improve individual documentation [^2][^3][^4].
- A final piece of missing strategy for smoke tests [^5].

All together we now have a complete picture of how we test GOV.UK,
which this new manual summarises.

This ability to describe how we test GOV.UK is the first step towards
making it better, as agreed in RFC 128 [^6]. The next step is to look
at the culture of testing we have in each team and agree what we need
to do to meet these new standards.

[^1]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-128-continuous-deployment.md#safety-checks
[^2]: https://github.com/alphagov/govuk-developer-docs/pull/3498
[^3]: https://github.com/alphagov/publishing-e2e-tests/pull/474
[^4]: https://github.com/alphagov/govuk_app_config/pull/244
[^5]: https://github.com/alphagov/smokey/pull/965
[^6]: https://github.com/alphagov/govuk-rfcs/blob/main/rfc-128-continuous-deployment.md#better-longterm-testing