Update dependencies owner for collections
Changes the dependency owner to match the application owner. The
coronavirus team doesn't exist anymore so the dependencies are not
being monitored.

It seems that if the `dependencies_team` is the same as the team,
doesn't need to be specified and can just be removed, so the entry for
local-links-manager has also been updated.