Reorder specialised machine reboot instructions alphabetically
It was hard to scan the list and quickly identify whether a given
machine class has special instructions for rebooting. Making the
list alphabetical should help with that.

I've also standardised the formatting, which now uses the actual
puppet node class name of the machine, i.e. `mongo` rather than
`MongoDB`, so that there is no ambiguity about what machine class
is being referred to.