Rename resync-mongo-db.html.md to mongo-db-commands.html.md
We're about to consolidate all MongoDB devops/infra commands in
one place, as they're currently spread around the Developer Docs,
leading to duplication and divergence. Consolidation should make
it easier to remove references to the (now deprecated) Fabric
scripts.

This file seems the most suitable starting point, as other docs
are either alert-related (so should be tightly scoped to just
dealing with said alert) or are only tangential to MongoDB (e.g.
the 'rebooting machines' doc).