Clarify when to create a non-tech leaver card
Previously this section said to create a tech leaver card
when the user was in the `govuk_tech.yml` and a non-tech
leaver card when they were in `govuk_non_tech.yml`. But it
didn't say anything when the user was in neither (which happens).

The Trello template ticket for leavers has the answer:
"you can confirm the leaver was in a tech role by finding their name in govuk_tech.yml. If they're not there, you may need this card: ... (non tech role).

This card updates the documentation to relect that.