Fixes to the 'restore RDS instance from snapshot' documentation
The previous iteration was geared towards how to drill the process.
This iteration expands on the instructions to account for what the process would be outside of a drill.

In writing this, I've arrived at the conclusion that the "manually getting back to where we were" steps are quite laborious/risky/error-prone, so documenting a Terraform approach would be better. I'm not quite sure what the implementation steps would be, but that can be documented later. Crucially, we have now documented the immediate steps to follow during an incident, so any tidy-up steps can be discussed post-incident and the docs updated. Hopefully the non-optimal route documented here is still a net improvement on what was there before.

NB I've also fixed a couple of broken headings and links.

https://trello.com/c/uqU7EHja/2949-figure-out-how-to-get-apps-to-point-to-new-database