Document change in process for getting Production Admin access
The old workflow expected access to be granted and revoked on the
very first shadow shift. Granting Production Admin access to all
the necessary systems requires a significant level of time and
effort, and it's clear the documented process was somewhat
idealistic and not rigidly followed.

After discussing with the 2nd line leads and the Lead Developers,
we've agreed to change the process such that no additional access
is granted on the first shadow shift. Access is now to be granted
on the second shadow shift (assuming probation has been passed),
and the access is retained by the developer at the end of the
shift. This means the work only has to be applied once, not three
times.

A concern was that the shadow developer may not be able to learn
much without the additional access, but there's already a
[sizeable collection of things to do](https://trello.com/c/MdGOVnLd/1210-shadowing-2nd-line)
which only Integration Admin access is required for.

Another consideration was whether the "Lead Developer / Head of
Technology" discretion rule was still necessary.
[It was applied recently](https://github.com/alphagov/govuk-aws-data/pull/1066),
so we've kept that breakglass mechanism in.

Finally, we've documented this all in a
[flow diagram](https://docs.google.com/presentation/d/10oRKrXqYki7LSFUySjb1e_FdTYGLihMKjeoH9qWC1kU/edit),
which should make these rules easier to parse.

As part of the docs cleanup, I've fixed the heading choice to
match the "When you get Production Deploy access" markup. I've
also removed the headings regarding "temporary" production admin
access, now that that concept no longer really exists.