Update ruby upgrade docs to reflect the current process
The PRs used to be generated by the Platform Reliability team. This will no
longer be the case and now the teams are solely responsible for upgrading ruby
versions in the apps/repos they own. Running the script with access token from
a personal account would unnecessarily increase the burden on team's devs.

The PRs would also be posted to the slack channel by the Seal. Since upgrading
to the latest ruby version doesn't tend to be the priority, the PRs are likely
to hang around for long time and add noise to the Slack channels.

It removes the example PR because it got out of date and the setup instructions
are now well described in the README.

Related PR: https://github.com/alphagov/upgrade-ruby-version/pull/8