Reorder sections in third party licensify access
These were in a bit of a weird order before, with things like "Accessing
MongoDB" coming before "Accessing machines using SSH" (even though accessing
machines using SSH is a prerequisite of accessing mongoDB.

I think starting from the simplest thing, it should be:

* Accessing the source code
* Accessing the logs
* Accessing the VPN
* Deploying code with Jenkins
* Acccessing machines using SSH
* Accessing MongoDB

Which is what I've gone for here.

The wording hasn't been changed in any of the sections.