Retire side-by-side-browser
This application is being retired. It has been non-operational for an
unclear amount of time due to being hosted on a personal Heroku account.
Due to this strange hosting it has never been a full part of the GOV.UK
stack.

We decided that, since to the very infrequent usage of transition, it
was a better option to retire the app that try to resurrect it.