Add policy on Ruby versions for Rubygems
This adds documentation on the approach we take for managing Ruby
versions with gems, this is added as documentation is this is not widely
known. I've also added some links to documentation that may be used when
updating Ruby to hopefully bring this to peoples attention (thanks for
the suggestions Murilo, Ollie and Jonathan).

With our Ruby apps we aim to run them, in dev, CI and production on
the newest Ruby version available. However with a Rubygem we aim for it
to be compatible across the range of supported Ruby versions - this
allows our apps to upgrade Ruby versions when the teams are ready and is
friendly to anyone outside GDS who also uses the gems.

When supporting a number of Ruby versions it is better to run on the
lowest version by default in development. This is because it immediately
prevents using features that aren't available in all of the supported
Rubies and it stops the linter trying to make suggestions to utilise
those new features.