Update docs for reordering role appointments
During a reshuffle, a minister can gain two positions of state, known as
RoleAppointments in Whitehall. One of these is perceived to be a more
senior role than the other. However, Collections will just render these
by ascending order of ActiveRecord IDs as default.

We've built a UI into Whitehall so we don't have to manually hit the
db via Rails console.

This updates the docs with instructions of how to access the UI.