Update Content Security Policy for new reporter
This does an overall update of our Content Security Policy
documentation, which hasn't really been changed since 2019. It's updated
to reflect the current status of the work and the somewhat long history.

We now have a new report system [1] which means we can now query reports
again, this includes documentation on how to query the reports.

[1]: https://github.com/alphagov/govuk-aws/pull/1664