Modernise docs around Reliability Engineering
SREs no longer have a separate rota, nor do they have a dedicated
Zendesk queue. SREs mostly (but not exclusively) work in the
Replatforming team. The official advice is to ask for help in
Platform Reliability, but until the two subteams are merged, it
seems helpful to link to the main Replatforming channel for now.

I've removed the reference to 'production' as it is self evident
later on in the documentation.