Make auto-merging workflow wait for tests to pass
Currently the auto-merging workflow doesn't work as it doesn't
have permission to merge to the repo, because not all PR checks
have passed at the time the workflow runs.

This change uses https://github.com/marketplace/actions/wait-on-check
to fix this.

I've tested the approach on a personal repo, here:
https://github.com/ChrisBAshton/test-auto-merge/pull/1

Trello: https://trello.com/c/uwoMBinS/3015-create-proof-of-concept-for-auto-merging-internal-prs-5