Make auto-merger wait on auto-approval of PR before merging
Currently, the auto-merger waits for CI to pass, before
attempting to merge. However, the merge fails, as PRs require
at least one approving review before merge.

We've already established the conditions under which an auto
review is sufficient: a patch upgrade of govuk_publishing_components,
where the associated CI checks pass. Therefore we already approve
of the PR, we just need a mechanism to make that happen.
We're using https://github.com/hmarr/auto-approve-action to
use a GitHub Actions 'bot' to approve the PR.

I've tested this approach in:
https://github.com/ChrisBAshton/test-auto-merge/pull/2

Trello: https://trello.com/c/uwoMBinS/3015-create-proof-of-concept-for-auto-merging-internal-prs-5