Change how Apps are grouped by Host name
After the Ruby 3.1.3 upgrade the following error was being returned when
the Apps page was built:

```
Error processing resource for index: apps.html
missing keywords: :label, :apps
 /home/runner/work/govuk-developer-docs/govuk-developer-docs/source/apps.html.md.erb:34:in `block in __tilt_19500'
 /home/runner/work/govuk-developer-docs/govuk-developer-docs/source/apps.html.md.erb:34:in `each'
 /home/runner/work/govuk-developer-docs/govuk-developer-docs/source/apps.html.md.erb:34:in `__tilt_19500'
```

This error only occurs for the "Apps by host" section.
It's not clear what has changed that is causing the
`Hosts::hosters_descending` method to fail. The method is only being
used in one place, and seems an overly complicated way of formatting the
host names.