Add PageVersions extension
This is a middleman extension to enable writing alternative versions of
a page. Page versions can be denoted by a particular attribute in the
pages frontmatter. Non-default pages are hidden from site navigation
features (such as search and list of pages), this is to prevent
the same page appearing more than once. The extension provides a helper
function to build page navigation to alternative versions.

This feature is being added to support alternative documentation for our
new EKS platform, whilst continuing to provide our existing
documentation. This allows readers to find documenation for the new
platform in a familiar place.