Activate PageVersions extension for platforms
This enable writing alternative versions of pages for different
platforms. This is helpful whilst we migrate to our new EKS platform. It
allows us to have both old and new existing documentation.

To write a alternative page version, pages need to specify the
`platform` attribute in frontmatter. At least one version should have
the value `EC2` as it's the default version.