Add more context to changing routes docs
When doing a drill without a real world scenario/context it was unclear what
this step was about.  Assuming the scenario was that someone was trying to
publish a document but got a route already taken error so after the route was
completely removed we are now republishing the document.